consolidated expandbox - no more expandbox2
